Moved/removed sqrtMat function
I removed the sqrtMat file because it is just one function used by
getRMDO.

So I moved that function to be a function within getRMDO() —
essentially a private function to getRMDO(). That means it cannot be
called outside of getRMDO (but that’s OK because it does not appear to
be called from anywhere else).